% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_app.R
\name{findviews_to_compare}
\alias{findviews_to_compare}
\title{Views of a multidimensional dataset, ranked by their differentiation power.}
\usage{
findviews_to_compare(group1, group2, data, view_size_max = NULL,
  clust_method = "complete", ...)
}
\arguments{
\item{group1}{Logical vector of size \code{nrow(data)}, which describes the
first group to compare. The value \code{TRUE} at position i indicates the
the i-th row of \code{data} belongs to the group.}

\item{group2}{Logical vector, which describes the second group to compare.
The value \code{TRUE} at position i indicates the the i-th row of
\code{data} belongs to the group.}

\item{data}{Data frame or matrix to be visualized}

\item{view_size_max}{Maximum number of columns in the views to be returned.
If this parameter is set to \code{NULL}, findviews uses
\code{log2(ncol(data)))}, rounded upwards if necessary.}

\item{clust_method}{Character describing a clustering method, used internally by
\code{\link[stats]{hclust}}. Examples are "complete", "single" or "average".}

\item{...}{Optional Shiny parameters, used in Shiny's
\code{\link[shiny]{runApp}} function.}

\item{nbins}{Number of bins used to discretize the target variable.}
}
\description{
\code{findviews_to_compare} detects views on which two arbitrary sets of
tuples are well separated. It plots the results with ggplot and Shiny.
}
\details{
The function \code{findviews_to_compare} takes two groups of tuples as input,
and detects views on which the statistical distribution of those two groups
is different.

To detect the views, \code{findviews_to_predict} concatenate the two groups
and applies the same method as \code{\link{findview}}.

To evaluate their differentiation power, it behaves differently according to
the type of data. To compare continuous data, it computes a synthetic score,
based on the difference between the means, the difference between the
variances and the difference between coefficient of of correlations. For
categorical data, it exploits the cosine distance between the histograms.

This method is roughly based on the following paper: \preformatted{
Fast, Explainable View Detection to Characterize Exploration Queries
Thibault Sellam, Martin Kersten
SSDBM, 2016}
}
\examples{
findviews_to_compare(mtcars$mpg >= 20 , mtcars$mpg < 20 , mtcars)

}

